+++ index.html
@@ -232,6 +232,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script defer src="./js/main.js"></script>
+  <script defer type="module" src="./js/main.js"></script>
 </body>
 </html>
